+++ src/Libs/CUDA_Utils/sources/CUDA_Utils.cu
@@ -8,12 +8,12 @@ template __global__ void axpy<uint64_t, double>(const double, const double*, dou
 template __global__ void axpy<uint32_t, __nv_bfloat16>(const __nv_bfloat16, const __nv_bfloat16*, __nv_bfloat16*, const uint32_t);
 template __global__ void axpy<uint64_t, __nv_bfloat16>(const __nv_bfloat16, const __nv_bfloat16*, __nv_bfloat16*, const uint64_t);
 
-template __global__ void dot_product<uint32_t, float>(const float*, const float*, float*, uint32_t);
-template __global__ void dot_product<uint64_t, float>(const float*, const float*, float*, uint64_t);
-template __global__ void dot_product<uint32_t, double>(const double*, const double*, float*, uint32_t);
-template __global__ void dot_product<uint64_t, double>(const double*, const double*, float*, uint64_t);
-template __global__ void dot_product<uint32_t, __nv_bfloat16>(const __nv_bfloat16*, const __nv_bfloat16*, float*, uint32_t);
-template __global__ void dot_product<uint64_t, __nv_bfloat16>(const __nv_bfloat16*, const __nv_bfloat16*, float*, uint64_t);
+template __global__ void dot_product<uint32_t, float>(const float*, const float*, double*, uint32_t);
+template __global__ void dot_product<uint64_t, float>(const float*, const float*, double*, uint64_t);
+template __global__ void dot_product<uint32_t, double>(const double*, const double*, double*, uint32_t);
+template __global__ void dot_product<uint64_t, double>(const double*, const double*, double*, uint64_t);
+template __global__ void dot_product<uint32_t, __nv_bfloat16>(const __nv_bfloat16*, const __nv_bfloat16*, double*, uint32_t);
+template __global__ void dot_product<uint64_t, __nv_bfloat16>(const __nv_bfloat16*, const __nv_bfloat16*, double*, uint64_t);
 
 template __global__ void scale<uint32_t, float>(const float, float*, uint32_t);
 template __global__ void scale<uint64_t, float>(const float, float*, uint64_t);
